make changes

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -18,7 +18,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 		return (NULL);
 
 	new_table->size = size;
-	new_table->array = calloc(size, sizeof(hash_node_t *));
+	new_table->array = calloc((size_t) size, sizeof(hash_node_t *));
 	if (new_table->array == NULL)
 		return (NULL);
